Refactored and renamed NeuralNet classes. Moved them to libEntendre so that they can be used in libNeat and other future libraries. This will require changes in the neural-net-tests repo.
