+++ libEntendre/SConscript
@@ -1,0 +1,7 @@
+Import('env')
+
+env.Append(CPPPATH=[Dir('include')])
+lib = env.SharedLibrary('Entendre',Glob('src/*.cc'),
+                        LIBS=[])
+
+Return('lib')
