+++ libNeat/SConscript
@@ -1,6 +1,7 @@
 Import('env')
 
 env.Append(CPPPATH=[Dir('include')])
+env.Append(CPPPATH=[Dir('../libEntendre/include')])
 lib = env.SharedLibrary('Neat',Glob('src/*.cc'),
                         LIBS=[])
