--- libNeat/src/NeuralNet.cc
@@ -1,164 +1,0 @@
-#include "NeuralNet.hh"
-
-#include <cassert>
-#include <stdexcept>
-#include <cmath>
-#include <iostream>
-#include <map>
-#include <set>
-#include <sstream>
-#include <vector>
-
-#include "Genome.hh"
-
-
-std::vector<double> NeuralNet::evaluate(std::vector<double> inputs) {
-  sort_connections();
-  load_input_vals(inputs);
-
-  for(auto& conn : connections) {
-    double input_val = get_node_val(conn.origin);
-    add_to_val(conn.dest, input_val * conn.weight);
-  }
-
-  return read_output_vals();
-}
-
-void NeuralNet::sort_connections() {
-  if(connections_sorted) {
-    return;
-  }
-
-  auto num_connections = connections.size();
-
-  std::vector<Connection> sorted;
-  sorted.reserve(num_connections);
-
-  std::vector<bool> used(num_connections,false);
-
-  for(size_t i = 0; i<num_connections; i++) {
-    size_t possible;
-    for(possible = 0; possible<num_connections; possible++) {
-
-      if (used[possible]) { continue; }
-
-      Connection& conn = connections[possible];
-      bool disqualified = false;
-
-      // Origin of normal connection has no unused input connections
-      if(conn.type == ConnectionType::Normal) {
-        for(size_t j=0; j<num_connections; j++) {
-          Connection& other = connections[j];
-          if(!used[j] &&
-             conn.origin == other.dest) {
-            disqualified = true;
-            break;
-          }
-        }
-      }
-      if(disqualified) {
-        continue;
-      }
-
-      // Destination of connection has no unused recurrent output connections
-      // If the output recurrent connection is ourself, it is allowed.
-      for(size_t j=0; j<num_connections; j++) {
-        Connection& other = connections[j];
-        if(!used[j] &&
-           conn.dest == other.origin &&
-           possible != j &&
-           other.type == ConnectionType::Recurrent) {
-          disqualified = true;
-          break;
-        }
-      }
-      if(disqualified) {
-        continue;
-      }
-
-      // Hurray, we found one!
-      break;
-    }
-
-    if(possible == num_connections) {
-      throw std::runtime_error("Sorting failed. Replace this with a class specific Exception");
-    }
-
-    used[possible] = true;
-    sorted.push_back(connections[possible]);
-  }
-
-  connections = sorted;
-  connections_sorted = true;
-}
-
-void LFNetwork::sort_connections() {
-
-  unsigned int max_iterations =
-    connections.size()*connections.size()*connections.size()+1;
-
-  bool change_applied = false;
-  for(unsigned int i_try=0; i_try < max_iterations; i_try++) {
-    change_applied = false;
-
-    for(int i=0; i<connections.size(); i++) {
-      for(int j=i+1; j<connections.size(); j++) {
-        Connection& conn1 = connections[i];
-        Connection& conn2 = connections[j];
-
-        switch(compare_connections(conn1,conn2)) {
-        case EvaluationOrder::GreaterThan:
-          if (conn1.set <= conn2.set) {
-            conn1.set = conn2.set + 1;
-            change_applied = true;
-          }
-          break;
-
-        case EvaluationOrder::LessThan:
-          if(conn2.set <= conn1.set) {
-            conn2.set = conn1.set + 1;
-            change_applied = true;
-          }
-          break;
-
-        case EvaluationOrder::NotEqual:
-          if(conn1.set == conn2.set) {
-            conn2.set = conn1.set + 1;
-            change_applied = true;
-          }
-          break;
-
-        case EvaluationOrder::Unknown:
-          break;
-        }
-      }
-    }
-
-    if(!change_applied) {
-      break;
-    }
-  }
-
-  if(change_applied) {
-    throw std::runtime_error("Sort Error: change_applied == true on last possible iteration");
-  }
-
-}
-
-
-LFNetwork::EvaluationOrder LFNetwork::compare_connections(const Connection& a, const Connection& b) {
-  if (a.type == ConnectionType::Recurrent && a.origin == b.dest) { return EvaluationOrder::LessThan; }
-  if (b.type == ConnectionType::Recurrent && b.origin == a.dest) { return EvaluationOrder::GreaterThan; }
-  if (a.type == ConnectionType::Normal && a.dest == b.origin) { return EvaluationOrder::LessThan; }
-  if (b.type == ConnectionType::Normal && b.dest == a.origin) { return EvaluationOrder::GreaterThan; }
-  if (a.dest == b.dest) { return EvaluationOrder::NotEqual; }
-  return EvaluationOrder::Unknown;
-}
-
-////////////////////////////////////////////////////////////////////////////
-
-LFNetwork::LFNetwork(NeuralNet&& net)
-  : NeuralNetBase(std::move(net.nodes),std::move(net.connections)) {
-  sigma = net.sigma;
-  connections_sorted = net.connections_sorted;
-}
